Fix the 2026-09-12 review findings; merge PRs #7-#11
Data safety
- partmove: TD64 high offset word went to iotd_Count instead of io_Actual;
  every MOVE/ZEROPART past 4 GB wrapped to offset mod 4 GB.  Move/zero now
  use the shared multi-block BlockDev API (also fixes MOVE/ZERO on images).
- sfsresize: bitmap bits were read/written through native ULONG loads
  (scrambled bitmaps on the Linux build) and the SWAR popcount used a 64-bit
  multiply on LP64 (garbage free-block counts).  Both byte-exact now.
- ADDPART BLOCKSIZE (CLI/script) set DE_SIZEBLOCK, physically overlapping
  neighbours; now DE_SECSPERBLK like the GUI.  Overlap checks (RDB_Write,
  CHECK, ADDMBR) compare physical spans.
- GUI FFS grow writes the RDB immediately (PFS/SFS/shrink already did).
- Truncated PART/FSHD chains: RDB_Read records the break, RDB_Write refuses
  until acknowledged (GUI requester, CLI/script warning + FORCE).
- RDB_Write writes the RDSK last and retires stale RDSK blocks in 0..15;
  BADB chains from the bad-block scan survive a Write.
- RESTORE writes the header to its own RDBBlocksLo (MBR+RDB disks);
  BACKUPEXT never saves less than the metadata seen; RESTOREEXT checks the
  backup belongs to the disk.
- Resize RDB write failure: one retry, then exact recovery instructions.
- Grow/shrink/DELPART unmount by device+unit, never by name alone.
- COPYDISK/IMAGEIN refuse a target with mounted partitions unless forced;
  copies report zero-filled unreadable blocks; PARTCLONE checks the SFS
  fixup result and commits destination geometry only on success.
- FS manager: Cancel keeps dirty after edits/deletes; delete no longer
  relabels partitions as FFS.  dol_Startup < 1024 guard (issue #12 shape).

Crashes / overflows
- Window titles and Raw Block Scan lines bounded; engine err_buf contract
  (ENGINE_ERRBUF_SIZE, snprintf everywhere); ask_yn no longer formats its
  buffer into itself.

Host build
- SetFileSize implemented; IMAGE CREATE works.  64-bit positioned file I/O:
  images beyond 2 GB.  Ctrl-C cancels copies cleanly.

Build / quality
- -Wall on, warning suppressions removed (GCC 14 needed gt_compat.h for
  Bartman's CreateContext macro); $VER string present and derived from
  version.h; dead Intuition includes and unused code removed.
- Read path remembers the working command; multi-block reads/writes in
  image copy and clone.

GUI / features
- Partition dialog lists FileSystem.resource filesystems (issue #13),
  validates names; Check FFS Root honours SECSPERBLK and the real root
  formula; CLI/script partition defaults match the GUI.

Tests / docs
- testloop.py verifies SFS bitmap contents; test/difftest.py compares
  m68k (vamos) and native results byte-for-byte (CI step added).
- cmdline.txt, FEATURES.txt, README, host/README updated; catalog UTF-8
  fixed and gencat.py warns about it.  REVIEW-2026-09-12.md records the
  review and the status of each finding.

+++ src/partview_shrink.c
@@ -57,5 +57,5 @@ void pv_shrink_report(struct Window *win, struct BlockDev *bd,
        percentage), which would be eaten if passed as the format itself. */
     es.es_TextFormat   = (UBYTE *)"%s";
     es.es_GadgetFormat = (UBYTE *)GS(MSG_OK);
-    EasyRequest(win, &es, NULL, buf);
+    EasyRequest(win, &es, NULL, (ULONG)buf);
 }
